requirements.txt
asgiref==3.2.10
chardet==2.3.0
Django==1.11.11
django-cors-headers==2.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.2.0
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
requests==2.9.1
six==1.10.0
sqlparse==0.3.1
urllib3==1.13.1
